Refactor change event logic into manager functions.
The code the maintains the list of changes is centralized into
a set of three functions. This gets rid of the wierd pattern wher
the count got passed down two levels so that the procesing functions
could increment it, but meanwhile the actual array was a global.

Fix a typo in a log statement string and a wrongly named call.

Made sure that "search" was not printed if there the search domain
list is empty.
.