RFC 6106 compliance
This change implements the procedure described in sections
6.2 and 6.3. Previously, the system kept only the latest
received RDNSS or DNSSL "batch", but the idea is that indivisual
entries in each list are to be managed separately, with their
own expire times. This change does this, and reworks the timer
management to accommodate it.

Also, some variable names were improved and there is now a
state dump triggered by sending the process a SIGUSR1 signal.
The man page has been updated to describe the state dump.